+++ crates/alef-scaffold/src/languages/go.rs
@@ -3,7 +3,7 @@ use alef_core::config::AlefConfig;
 use alef_core::ir::ApiSurface;
 use std::path::PathBuf;
 
-pub fn scaffold_go(api: &ApiSurface, config: &AlefConfig) -> anyhow::Result<Vec<GeneratedFile>> {
+pub(crate) fn scaffold_go(api: &ApiSurface, config: &AlefConfig) -> anyhow::Result<Vec<GeneratedFile>> {
     let go_module = config.go_module();
     let version = &api.version;
     let _ = version; // go.mod doesn't embed the package version
